+++ isy-serviceapi-core/src/main/java/de/bund/bva/isyfact/serviceapi/core/httpinvoker/TimeoutWiederholungHttpInvokerRequestExecutor.java
@@ -21,39 +21,47 @@ import java.io.IOException;
 import java.io.InterruptedIOException;
 import java.net.HttpURLConnection;
 
+import org.springframework.http.HttpHeaders;
 import org.springframework.remoting.httpinvoker.HttpInvokerClientConfiguration;
 import org.springframework.remoting.httpinvoker.SimpleHttpInvokerRequestExecutor;
 import org.springframework.remoting.support.RemoteInvocationResult;
 
+import de.bund.bva.isyfact.aufrufkontext.AufrufKontextVerwalter;
 import de.bund.bva.isyfact.logging.IsyLogger;
 import de.bund.bva.isyfact.logging.IsyLoggerFactory;
 import de.bund.bva.isyfact.logging.LogKategorie;
 import de.bund.bva.isyfact.serviceapi.common.konstanten.EreignisSchluessel;
 
 /**
- * Erweiterung des {@link SimpleHttpInvokerRequestExecutor} von Spring. Diese Erweiterung erlaubt es den
- * Timeout und eine Aufrufwiederholung zu konfigurieren.
- *
- *
+ * Extension to the Spring {@link SimpleHttpInvokerRequestExecutor} which allows
+ * to configure the timeout and the times a call is repeated.
  */
 public class TimeoutWiederholungHttpInvokerRequestExecutor extends SimpleHttpInvokerRequestExecutor {
 
     /** Isy-Logger. */
     private static final IsyLogger LOG = IsyLoggerFactory
-        .getLogger(TimeoutWiederholungHttpInvokerRequestExecutor.class);
+            .getLogger(TimeoutWiederholungHttpInvokerRequestExecutor.class);
 
-    /** Timeout für Request. */
+    /** {@link AufrufKontextVerwalter} to set the OAuth 2 Bearer Token. */
+    private final AufrufKontextVerwalter<?> aufrufKontextVerwalter;
+
+    /** Timeout for the request. */
     private int timeout;
 
-    /** Anzahl Wiederholungen bei Timeouts. */
+    /** Number of times the call is repeated in case of a timeout. */
     private int anzahlWiederholungen;
 
-    /** Pause zwischen Wiederholungen. */
+    /** Break between the call repetitions. */
     private int wiederholungenAbstand;
 
     /**
-     * {@inheritDoc}
+     * Constructor. Sets the {@link AufrufKontextVerwalter} bean.
+     * @param aufrufKontextVerwalter new value for the {@link AufrufKontextVerwalter}.
      */
+    public TimeoutWiederholungHttpInvokerRequestExecutor(AufrufKontextVerwalter<?> aufrufKontextVerwalter) {
+        this.aufrufKontextVerwalter = aufrufKontextVerwalter;
+    }
+
     @Override
     protected RemoteInvocationResult doExecuteRequest(HttpInvokerClientConfiguration config,
         ByteArrayOutputStream baos) throws IOException, ClassNotFoundException {
@@ -65,16 +73,16 @@ public class TimeoutWiederholungHttpInvokerRequestExecutor extends SimpleHttpInv
                 LOG.info(LogKategorie.PROFILING, EreignisSchluessel.TIMEOUT,
                     "Beim Aufrufen des Services [{}] ist ein Timeout aufgetreten.", config.getServiceUrl());
                 versuch++;
-                if (versuch == this.anzahlWiederholungen) {
+                if (versuch == anzahlWiederholungen) {
                     LOG.info(LogKategorie.PROFILING, EreignisSchluessel.TIMEOUT_ABBRUCH,
                         "Aufruf nach Timeout abgebrochen.");
                     throw requestException;
                 }
                 try {
-                    if (this.wiederholungenAbstand > 0) {
+                    if (wiederholungenAbstand > 0) {
                         LOG.info(LogKategorie.PROFILING, EreignisSchluessel.TIMEOUT_WARTEZEIT,
-                            "Warte {}ms bis zur Wiederholung des Aufrufs.", this.wiederholungenAbstand);
-                        Thread.sleep(this.wiederholungenAbstand);
+                            "Warte {}ms bis zur Wiederholung des Aufrufs.", wiederholungenAbstand);
+                        Thread.sleep(wiederholungenAbstand);
                     }
                 } catch (InterruptedException ex) {
                     LOG.info(LogKategorie.PROFILING, EreignisSchluessel.TIMEOUT_WARTEZEIT_ABBRUCH,
@@ -87,43 +95,44 @@ public class TimeoutWiederholungHttpInvokerRequestExecutor extends SimpleHttpInv
         }
     }
 
-    /**
-     * {@inheritDoc}
-     */
     @Override
     protected void prepareConnection(HttpURLConnection con, int contentLength) throws IOException {
         super.prepareConnection(con, contentLength);
-        con.setReadTimeout(this.timeout);
-        con.setConnectTimeout(this.timeout);
+        if (aufrufKontextVerwalter.getBearerToken() != null) {
+            con.setRequestProperty(HttpHeaders.AUTHORIZATION, "Bearer " + aufrufKontextVerwalter.getBearerToken());
+        } else {
+            LOG.info(LogKategorie.JOURNAL, EreignisSchluessel.KEIN_BEARER_TOKEN_IM_AUFRUFKONTEXT,
+                    "Kein Bearer-Token im AufrufKontextVerwalter. Der Authorization-Header wird nicht gesetzt.");
+        }
+        con.setReadTimeout(timeout);
+        con.setConnectTimeout(timeout);
     }
 
     /**
-     * Setzt den Timeout in Millisekunden. Der Timeout wird beim Aufbau und beim Lesen über die
-     * HTTP-Connection verwendet.
+     * Sets the timeout in milliseconds. The timeout is used when establishing and reading via the HTTP connection.
      * @see HttpURLConnection#setConnectTimeout(int)
      * @see HttpURLConnection#setReadTimeout(int)
      * @param timeout
-     *            Timeout in Millisekunden.
+     *            timeout in milliseconds.
      */
     public void setTimeout(int timeout) {
         this.timeout = timeout;
     }
 
     /**
-     * Hierüber wird festgelegt, wie oft der Aufruf bei einem Timeout wiederholt werden soll. Default ist 0.
+     * This defines how often the call is to be repeated in the event of a timeout. Default value is 0.
      * @param anzahlWiederholungen
-     *            Anzahl Wiederholungen bei Timeouts.
+     *            number of repetitions in case of a timeout.
      */
     public void setAnzahlWiederholungen(int anzahlWiederholungen) {
         this.anzahlWiederholungen = anzahlWiederholungen;
     }
 
     /**
-     * Hierüber wird festgelegt, wie lange zwischen zwei Aufrufwiederholungen gewartet werden soll. Default
-     * ist 0.
+     * This defines how long to wait between two call repetitions. Default value is 0.
      * @see #setAnzahlWiederholungen(int)
      * @param wiederholungenAbstand
-     *            Pause zwischen den Wiederholungen in Millisekunden.
+     *            break between repetitions in milliseconds.
      */
     public void setWiederholungenAbstand(int wiederholungenAbstand) {
         this.wiederholungenAbstand = wiederholungenAbstand;