+++ isy-serviceapi-core/CHANGELOG.md
@@ -1,4 +1,7 @@
 # 2.2.0
+- `IFS-874`: Umsetzung des Transports von OAuth 2 Bearer Tokens zwischen Schnittstellentechnologien
+    - `IsyHttpInvokerServiceExporter` und `TimeoutWiederholungHttpInvokerRequestExecutor` um den `AufrufKontextVerwalter` erweitert.
+    - Der OAuth 2 Bearer Token wird über den HTTP Authentication Header transportiert bzw. aus diesem ausgelesen.
 - `IFS-564`: Deaktivieren von Proxy-Objekten in HttpInvoker Schnittstellen
 - `IFS-597`: Verwendung der Klasse `IsyHttpInvokerServiceExporter` anstatt `HttpInvokerServiceExporter`
   aus Tests entfernt